requirements.txt
Django
djangorestframework
django-cors-headers
djoser
drf-yasg
Pillow
setuptools
beautifulsoup4
func-timeout